+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,15 @@
         android:theme="@style/Theme.TripAlbum"
         android:usesCleartextTraffic="true" >
         <activity
+            android:name=".view.FriendsListView"
+            android:exported="false" />
+        <activity
+            android:name=".view.SearchPlacesView"
+            android:exported="false" />
+        <activity
+            android:name=".view.MyAlbumView"
+            android:exported="false" />
+        <activity
             android:name=".view.ViewPlaceActivityView"
             android:exported="false" />
 
@@ -23,7 +32,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".view.ModifyVisitActivityView"
+            android:name=".view.NewVisitView"
             android:exported="false" />
         <activity
             android:name=".view.NewPlaceActivityView"
@@ -33,6 +42,9 @@
             android:exported="false" />
         <activity
             android:name=".view.NewProvinceActivityView"
+            android:exported="false" />
+        <activity
+            android:name=".view.ViewVisitsView"
             android:exported="false" />
         <activity
             android:name=".view.MainActivityView"
