+++ app/src/main/java/com/svalero/tripalbum/view/FriendsListView.java
@@ -1,0 +1,16 @@
+package com.svalero.tripalbum.view;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.svalero.tripalbum.R;
+
+public class FriendsListView extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_friends_list_view);
+    }
+}
