+++ app/src/main/res/layout-land/activity_new_visit.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".view.ModifyVisitActivityView">
+    tools:context=".view.NewVisitView">
 
     <LinearLayout
         android:layout_width="match_parent"
