+++ app/src/main/res/layout-land/activity_view_visits.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".view.MainActivityView">
+    tools:context=".view.ViewVisitsView">
 
     <LinearLayout
         android:layout_width="match_parent"
