+++ app/src/main/res/layout/activity_main.xml
@@ -4,111 +4,82 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:onClick="openViewVisits"
     android:padding="10dp"
     tools:context=".view.MainActivityView">
 
     <LinearLayout
-        android:id="@+id/linearLayout4"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:orientation="vertical"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.157">
+        app:layout_constraintTop_toTopOf="parent">
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="50dp"
+            android:layout_height="200dp"
             android:orientation="horizontal">
 
-            <TextView
-                android:id="@+id/textView13"
-                android:layout_width="0dp"
+            <Button
+                android:id="@+id/button_main_1"
+                android:layout_width="100dp"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:text="@string/country"
-                android:textSize="16sp" />
+                android:backgroundTint="@android:color/holo_red_dark"
+                android:insetTop="0dp"
+                android:insetBottom="0dp"
+                android:onClick="openMyAlbum"
+                android:text="@string/main_my_album" />
 
-            <ListView
-                android:id="@+id/countries_list_main"
-                android:layout_width="200dp"
+            <Button
+                android:id="@+id/button_main_2"
+                android:layout_width="100dp"
                 android:layout_height="match_parent"
-                android:layout_weight="1" />
+                android:layout_weight="1"
+                android:backgroundTint="@android:color/holo_orange_light"
+                android:insetTop="0dp"
+                android:insetBottom="0dp"
+                android:onClick="openFriendsList"
+                android:text="@string/main_friends" />
 
         </LinearLayout>
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="50dp"
+            android:layout_height="200dp"
             android:orientation="horizontal">
 
-            <TextView
-                android:id="@+id/textView14"
-                android:layout_width="0dp"
+            <Button
+                android:id="@+id/button_main_3"
+                android:layout_width="100dp"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:text="@string/province"
-                android:textSize="16sp" />
+                android:backgroundTint="@android:color/holo_blue_dark"
+                android:insetTop="0dp"
+                android:insetBottom="0dp"
+                android:onClick="openSearchPlaces"
+                android:text="@string/main_places" />
 
-            <ListView
-                android:id="@+id/provinces_list_main"
-                android:layout_width="200dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1" />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/textView15"
-                android:layout_width="0dp"
+            <Button
+                android:id="@+id/button_main_4"
+                android:layout_width="100dp"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:text="@string/place"
-                android:textSize="16sp" />
-
-            <ListView
-                android:id="@+id/places_list_main"
-                android:layout_width="200dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1" />
+                android:backgroundTint="@color/purple_700"
+                android:insetTop="0dp"
+                android:insetBottom="0dp"
+                android:onClick="openNewVisit"
+                android:text="@string/main_add_visit" />
         </LinearLayout>
 
-        <ListView
-            android:id="@+id/visit_list_main"
+        <Button
+            android:id="@+id/button2"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
+            android:layout_height="wrap_content"
+            android:onClick="openViewVisits"
+            android:text="View visits" />
 
     </LinearLayout>
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/view_place"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:clickable="true"
-        android:onClick="openViewPlace"
-        app:layout_constraintBottom_toTopOf="@+id/add_visit"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:srcCompat="@android:drawable/ic_dialog_map" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/add_visit"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:clickable="true"
-        android:onClick="openNewVisit"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:srcCompat="@android:drawable/ic_input_add" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
